* Add `force_sync_delete_ops` configuration parameter: the user should explicitly set it to true to allow networking-oneview to delete networks and ports due to changes in configuration mappings (these were enabled by default but as they can be dangerous and power off servers to make port changes we understand this is risky and should be a cloud administrator choice to enable them).
* Deprecate `developer_mode` configuration parameter

+++ networking_oneview/ml2/drivers/oneview/synchronization.py
@@ -56,9 +56,14 @@ class Synchronization(object):
     def synchronize(self):
         self.delete_outdated_flat_mapped_networks()
         self.create_oneview_networks_from_neutron()
-        self.delete_unmapped_oneview_networks()
-        self.synchronize_uplinkset_from_mapped_networks()
-        self.recreate_connection()
+
+        force_delete = common.CONF.DEFAULT.force_sync_delete_ops
+        LOG.debug("Delete outdated networks and connections operations "
+                  "is set to: %s" % force_delete)
+        if force_delete:
+            self.delete_unmapped_oneview_networks()
+            self.synchronize_uplinkset_from_mapped_networks()
+            self.recreate_connection()
 
     def get_oneview_network(self, oneview_net_id):
         try:
